requirements.txt
langchain
langchain-community
langchain-groq
langgraph
chromadb
pypdf
tiktoken
python-dotenv
sentence-transformers
streamlit